requirements.txt
# run this in terminal to install all required modules
# pip install -r /path/to/requirements.txt

# required modules
pandas
numpy
glob
os
matplotlib.pyplot
cv2
seaborn
sklearn.metrics
tensorflow
keras_tuner